+++ default.nix
@@ -1,6 +1,6 @@
 { mkDerivation, base, containers, directory, filepath, hashable
-, haskeline, mtl, parsec, stdenv, tasty, tasty-hunit, text
-, unordered-containers, wl-pprint
+, haskeline, mtl, parsec, stdenv, tasty, tasty-hunit
+, template-haskell, terminfo, text, unordered-containers, wl-pprint
 }:
 mkDerivation {
   pname = "expresso";
@@ -9,12 +9,16 @@ mkDerivation {
   isLibrary = true;
   isExecutable = true;
   libraryHaskellDepends = [
-    base containers directory filepath hashable mtl parsec text
-    unordered-containers wl-pprint
+    base containers directory filepath hashable haskeline mtl parsec
+    template-haskell terminfo text unordered-containers wl-pprint
+  ];
+  executableHaskellDepends = [
+    base containers directory filepath hashable haskeline mtl parsec
+    terminfo text unordered-containers wl-pprint
   ];
   testHaskellDepends = [
     base containers directory filepath hashable haskeline mtl parsec
-    tasty tasty-hunit text unordered-containers wl-pprint
+    tasty tasty-hunit terminfo text unordered-containers wl-pprint
   ];
   description = "A simple expressions language based on row types";
   license = stdenv.lib.licenses.bsd3;